% Generated by roxygen2 (4.0.0): do not edit by hand
\name{accplot}
\alias{accplot}
\title{Plot x, y, z, accelerometer data w.r.t. time}
\usage{
accplot(dir, x = TRUE, y = TRUE, z = TRUE)
}
\arguments{
\item{dir}{location of the bin file to be read}

\item{x}{boolean of whether x is to be plotted}

\item{y}{boolean of whether y is to be plotted}

\item{z}{boolean of whether z is to be plotted}
}
\description{
This function generates a plot of accelerometer data
w.r.t. time using ggplot. Options for selecting which
variables (x, y, or z) are to be plotted are available.
}

